refactor: structの関数化を行った
