+++ pkg/cruds/create/add_player.go
@@ -2,31 +2,14 @@ package createDB
 
 import (
 	connectDB "gamename-back-end/pkg/connect_db"
+	structManege "gamename-back-end/pkg/struct"
 	"log"
 
 	"cloud.google.com/go/firestore"
 )
 
-type Player struct {
-	PlayerName string
-	Icon       int
-	Role       int
-	Theme      string
-	Hint       string
-	IsDelete   bool
-	Answer     string
-	Wolf       bool
-	Vote       int
-	Point      int
-}
-
-type RoomPlayer struct {
-	RoomId   string
-	PlayerId string
-}
-
 func AddPlayer(roomId string, playerName string, playerIcon int) string {
-	player := Player{
+	player := structManege.CreatePlayer{
 		PlayerName: playerName,
 		Icon:       playerIcon,
 		Role:       0,
@@ -54,7 +37,7 @@ func AddPlayer(roomId string, playerName string, playerIcon int) string {
 	if err != nil {
 		log.Printf("An error has occurred: %s", err)
 	}
-	roomPlayer := RoomPlayer{
+	roomPlayer := structManege.RoomPlayer{
 		RoomId:   roomId,
 		PlayerId: docRef.ID,
 	}
